move in tree
