--- src/fakes/ray/ipc/BUILD.bazel
@@ -1,9 +1,0 @@
-load("//bazel:ray.bzl", "ray_cc_library")
-
-ray_cc_library(
-    name = "fake_raylet_ipc_client",
-    hdrs = ["raylet_ipc_client.h"],
-    deps = [
-        "//src/ray/ipc:raylet_ipc_client_interface",
-    ],
-)
