+++ src/ray/core_worker/test/BUILD.bazel
@@ -226,7 +226,7 @@ ray_cc_test(
     deps = [
         "//:ray_fakes",
         "//:ray_mock",
-        "//src/fakes/ray/ipc:fake_raylet_ipc_client",
+        "//src/ray/ipc:fake_raylet_ipc_client",
         "//src/ray/common:test_util",
         "//src/ray/core_worker:core_worker_lib",
         "//src/ray/core_worker:memory_store",
