+++ src/ray/ipc/BUILD.bazel
@@ -1,19 +1,24 @@
 load("//bazel:ray.bzl", "ray_cc_library")
 
 ray_cc_library(
-    name = "client_connection",
-    srcs = [
-        "client_connection.cc",
-    ],
-    hdrs = [
-        "client_connection.h",
-    ],
+    name = "raylet_ipc_client_interface",
+    hdrs = ["raylet_ipc_client_interface.h"],
     deps = [
-        "//src/ray/common:asio",
-        "//src/ray/common:event_stats",
+        "//src/ray/common:buffer",
         "//src/ray/common:id",
         "//src/ray/common:status",
         "//src/ray/flatbuffers:node_manager_generated",
+        "//src/ray/protobuf:common_cc_proto",
+        "//src/ray/util:process",
+        "@com_google_absl//absl/container:flat_hash_set",
+    ],
+)
+
+ray_cc_library(
+    name = "fake_raylet_ipc_client",
+    hdrs = ["raylet_ipc_client.h"],
+    deps = [
+        "//src/ray/ipc:raylet_ipc_client_interface",
     ],
 )
 
@@ -38,15 +43,18 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "raylet_ipc_client_interface",
-    hdrs = ["raylet_ipc_client_interface.h"],
+    name = "client_connection",
+    srcs = [
+        "client_connection.cc",
+    ],
+    hdrs = [
+        "client_connection.h",
+    ],
     deps = [
-        "//src/ray/common:buffer",
+        "//src/ray/common:asio",
+        "//src/ray/common:event_stats",
         "//src/ray/common:id",
         "//src/ray/common:status",
         "//src/ray/flatbuffers:node_manager_generated",
-        "//src/ray/protobuf:common_cc_proto",
-        "//src/ray/util:process",
-        "@com_google_absl//absl/container:flat_hash_set",
     ],
 )
